+++ CHANGELOG.md
@@ -2,12 +2,16 @@
 
 # UNRELEASED
 
+* EXOCASMailboxSettings
+  * New resource to configure Exchange Online CAS Mailbox settings.
 * IntuneAppProtectionPolicyiOS
    * Fixed issue with creation a new policies where it was complaining about invalid minimum versions.
    * Fixed issues where creating new policies threw an error complaining about an invalid duration format.
      FIXES [#2019](https://github.com/microsoft/Microsoft365DSC/issues/2019)
    * Added the CustomBrowserProtocol paramters.
      FIXES [#2009](https://github.com/microsoft/Microsoft365DSC/issues/2009)
+* IntuneDeviceAndAppManagementAssignmentFilter
+  * Initial release.
 * SCComplianceTag
   * Fixed issue where FilePlanProperty was not properly applied unless another child property was also modified.
 * DEPENDENCIES